Exception handling

+++ src/main/java/com/example/course/services/UserService.java
@@ -2,7 +2,12 @@ package com.example.course.services;
 
 import com.example.course.entities.User;
 import com.example.course.repositories.UserRepository;
+import com.example.course.services.exceptions.DataBaseException;
+import com.example.course.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -21,7 +26,7 @@ public class UserService {
     public User findById(Long id) {
         Optional<User> optionalUser = userRepository.findById(id);
 
-        return optionalUser.get();
+        return optionalUser.orElseThrow(() -> new ResourceNotFoundException(id));
     }
 
     public User insert(User user) {
@@ -29,15 +34,26 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        userRepository.deleteById(id);
+        try {
+            if (!userRepository.existsById(id)) {
+                throw new ResourceNotFoundException(id);
+            }
+            userRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DataBaseException(e.getMessage());
+        }
     }
 
+
     public User update(Long id, User updatedUser) {
-        User entity = userRepository.getReferenceById(id);
+        try{
+            User entity = userRepository.getReferenceById(id);
+            updateData(entity, updatedUser);
 
-        updateData(entity, updatedUser);
-
-        return userRepository.save(entity);
+            return userRepository.save(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException(id);
+        }
     }
 
     private void updateData(User entity, User updatedUser) {
